--- husky_base/husky_base.xml
@@ -1,10 +1,0 @@
-<library path="husky_base">
-
-  <class name="husky_base/HuskyBase"
-         type="husky_base::HuskyBase"
-         base_class_type="hardware_interface::SystemInterface">
-    <description>
-      The ros2_control HuskyBase using a system hardware interface-type. It uses velocity command and position state interface. The example is the starting point to implement a hardware interface for differential-drive mobile robots.
-    </description>
-  </class>
-</library>
