+++ husky_base/package.xml
@@ -13,13 +13,15 @@
 
   <buildtool_depend>ament_cmake</buildtool_depend>
 
+  <depend>controller_interface</depend>
+  <depend>controller_manager</depend>
+  <depend>controller_manager_msgs</depend>
   <depend>hardware_interface</depend>
-  <depend>pluginlib</depend>
-  <depend>rclcpp</depend>
-
   <depend>geometry_msgs</depend>
   <depend>husky_msgs</depend>
   <depend>nav_msgs</depend>
+  <depend>pluginlib</depend>
+  <depend>rclcpp</depend>
   <depend>sensor_msgs</depend>
   <depend>std_msgs</depend>
   <depend>std_srvs</depend>
