+++ husky_base/src/husky_hardware.cpp
@@ -29,187 +29,73 @@
 *
 */
 
-#include "husky_base/husky_hardware.h"
-#include <boost/assign/list_of.hpp>
+#include "husky_base/husky_hardware.hpp"
+
+#include <chrono>
+#include <cmath>
+#include <limits>
+#include <memory>
+#include <vector>
+
+#include "hardware_interface/types/hardware_interface_type_values.hpp"
+#include "rclcpp/rclcpp.hpp"
 
 namespace
 {
   const uint8_t LEFT = 0, RIGHT = 1;
-};
+}
 
 namespace husky_base
 {
-
-  /**
-  * Initialize Husky hardware
-  */
-  HuskyHardware::HuskyHardware(ros::NodeHandle nh, ros::NodeHandle private_nh, double target_control_freq)
-    :
-    nh_(nh),
-    private_nh_(private_nh),
-    system_status_task_(husky_status_msg_),
-    power_status_task_(husky_status_msg_),
-    safety_status_task_(husky_status_msg_),
-    software_status_task_(husky_status_msg_, target_control_freq)
-  {
-    private_nh_.param<double>("wheel_diameter", wheel_diameter_, 0.3302);
-    private_nh_.param<double>("max_accel", max_accel_, 5.0);
-    private_nh_.param<double>("max_speed", max_speed_, 1.0);
-    private_nh_.param<double>("polling_timeout_", polling_timeout_, 10.0);
-
-    std::string port;
-    private_nh_.param<std::string>("port", port, "/dev/prolific");
-
-    horizon_legacy::connect(port);
-    horizon_legacy::configureLimits(max_speed_, max_accel_);
-    resetTravelOffset();
-    initializeDiagnostics();
-    registerControlInterfaces();
-  }
+  static const std::string HW_NAME = "HuskyHardware";
 
   /**
   * Get current encoder travel offsets from MCU and bias future encoder readings against them
   */
   void HuskyHardware::resetTravelOffset()
   {
-    horizon_legacy::Channel<clearpath::DataEncoders>::Ptr enc = horizon_legacy::Channel<clearpath::DataEncoders>::requestData(
-      polling_timeout_);
+    horizon_legacy::Channel<clearpath::DataEncoders>::Ptr enc =
+        horizon_legacy::Channel<clearpath::DataEncoders>::requestData(polling_timeout_);
     if (enc)
     {
-      for (int i = 0; i < 4; i++)
+      for (auto i = 0u; i < hw_states_position_offset_.size(); i++)
       {
-        joints_[i].position_offset = linearToAngular(enc->getTravel(i % 2));
+        hw_states_position_offset_[i] = linearToAngular(enc->getTravel(i % 2));
       }
     }
     else
     {
-      ROS_ERROR("Could not get encoder data to calibrate travel offset");
+      RCLCPP_FATAL(
+        rclcpp::get_logger(HW_NAME), "Could not get encoder data to calibrate travel offset");
     }
   }
 
   /**
-  * Register diagnostic tasks with updater class
+  * Husky reports travel in metres, need radians for ros_control RobotHW
   */
-  void HuskyHardware::initializeDiagnostics()
+  double HuskyHardware::linearToAngular(const double &travel) const
   {
-    horizon_legacy::Channel<clearpath::DataPlatformInfo>::Ptr info =
-      horizon_legacy::Channel<clearpath::DataPlatformInfo>::requestData(polling_timeout_);
-    std::ostringstream hardware_id_stream;
-    hardware_id_stream << "Husky " << info->getModel() << "-" << info->getSerial();
-
-    diagnostic_updater_.setHardwareID(hardware_id_stream.str());
-    diagnostic_updater_.add(system_status_task_);
-    diagnostic_updater_.add(power_status_task_);
-    diagnostic_updater_.add(safety_status_task_);
-    diagnostic_updater_.add(software_status_task_);
-    diagnostic_publisher_ = nh_.advertise<husky_msgs::HuskyStatus>("status", 10);
-  }
-
-
-  /**
-  * Register interfaces with the RobotHW interface manager, allowing ros_control operation
-  */
-  void HuskyHardware::registerControlInterfaces()
-  {
-    ros::V_string joint_names = boost::assign::list_of("front_left_wheel")
-      ("front_right_wheel")("rear_left_wheel")("rear_right_wheel");
-    for (unsigned int i = 0; i < joint_names.size(); i++)
-    {
-      hardware_interface::JointStateHandle joint_state_handle(joint_names[i],
-                                                              &joints_[i].position, &joints_[i].velocity,
-                                                              &joints_[i].effort);
-      joint_state_interface_.registerHandle(joint_state_handle);
-
-      hardware_interface::JointHandle joint_handle(
-        joint_state_handle, &joints_[i].velocity_command);
-      velocity_joint_interface_.registerHandle(joint_handle);
-    }
-    registerInterface(&joint_state_interface_);
-    registerInterface(&velocity_joint_interface_);
+    return (travel / wheel_diameter_ * 2.0f);
   }
 
   /**
-  * External hook to trigger diagnostic update
+  * RobotHW provides velocity command in rad/s, Husky needs m/s,
   */
-  void HuskyHardware::updateDiagnostics()
+  double HuskyHardware::angularToLinear(const double &angle) const
   {
-    diagnostic_updater_.force_update();
-    husky_status_msg_.header.stamp = ros::Time::now();
-    diagnostic_publisher_.publish(husky_status_msg_);
+    return (angle * wheel_diameter_ / 2.0f);
   }
 
-  /**
-  * Pull latest speed and travel measurements from MCU, and store in joint structure for ros_control
-  */
-  void HuskyHardware::updateJointsFromHardware()
-  {
-
-    horizon_legacy::Channel<clearpath::DataEncoders>::Ptr enc = horizon_legacy::Channel<clearpath::DataEncoders>::requestData(
-      polling_timeout_);
-    if (enc)
-    {
-      ROS_DEBUG_STREAM("Received travel information (L:" << enc->getTravel(LEFT) << " R:" << enc->getTravel(RIGHT) << ")");
-      for (int i = 0; i < 4; i++)
-      {
-        double delta = linearToAngular(enc->getTravel(i % 2)) - joints_[i].position - joints_[i].position_offset;
-
-        // detect suspiciously large readings, possibly from encoder rollover
-        if (std::abs(delta) < 1.0)
-        {
-          joints_[i].position += delta;
-        }
-        else
-        {
-          // suspicious! drop this measurement and update the offset for subsequent readings
-          joints_[i].position_offset += delta;
-          ROS_DEBUG("Dropping overflow measurement from encoder");
-        }
-      }
-    }
-
-    horizon_legacy::Channel<clearpath::DataDifferentialSpeed>::Ptr speed = horizon_legacy::Channel<clearpath::DataDifferentialSpeed>::requestData(
-      polling_timeout_);
-    if (speed)
-    {
-      ROS_DEBUG_STREAM("Received linear speed information (L:" << speed->getLeftSpeed() << " R:" << speed->getRightSpeed() << ")");
-      for (int i = 0; i < 4; i++)
-      {
-        if (i % 2 == LEFT)
-        {
-          joints_[i].velocity = linearToAngular(speed->getLeftSpeed());
-        }
-        else
-        { // assume RIGHT
-          joints_[i].velocity = linearToAngular(speed->getRightSpeed());
-        }
-      }
-    }
-  }
-
-  /**
-  * Get latest velocity commands from ros_control via joint structure, and send to MCU
-  */
   void HuskyHardware::writeCommandsToHardware()
   {
-    double diff_speed_left = angularToLinear(joints_[LEFT].velocity_command);
-    double diff_speed_right = angularToLinear(joints_[RIGHT].velocity_command);
+    double diff_speed_left = angularToLinear(hw_commands_[LEFT]);
+    double diff_speed_right = angularToLinear(hw_commands_[RIGHT]);
 
     limitDifferentialSpeed(diff_speed_left, diff_speed_right);
 
     horizon_legacy::controlSpeed(diff_speed_left, diff_speed_right, max_accel_, max_accel_);
   }
 
-  /**
-  * Update diagnostics with control loop timing information
-  */
-  void HuskyHardware::reportLoopDuration(const ros::Duration &duration)
-  {
-    software_status_task_.updateControlFrequency(1 / duration.toSec());
-  }
-
-  /**
-  * Scale left and right speed outputs to maintain ros_control's desired trajectory without saturating the outputs
-  */
   void HuskyHardware::limitDifferentialSpeed(double &diff_speed_left, double &diff_speed_right)
   {
     double large_speed = std::max(std::abs(diff_speed_left), std::abs(diff_speed_right));
@@ -221,21 +107,249 @@ namespace husky_base
     }
   }
 
-  /**
-  * Husky reports travel in metres, need radians for ros_control RobotHW
-  */
-  double HuskyHardware::linearToAngular(const double &travel) const
-  {
-    return travel / wheel_diameter_ * 2;
-  }
 
   /**
-  * RobotHW provides velocity command in rad/s, Husky needs m/s,
+  * Pull latest speed and travel measurements from MCU, and store in joint structure for ros_control
   */
-  double HuskyHardware::angularToLinear(const double &angle) const
+  void HuskyHardware::updateJointsFromHardware()
   {
-    return angle * wheel_diameter_ / 2;
+
+    horizon_legacy::Channel<clearpath::DataEncoders>::Ptr enc =
+      horizon_legacy::Channel<clearpath::DataEncoders>::requestData(polling_timeout_);
+    if (enc)
+    {
+      RCLCPP_DEBUG(
+        rclcpp::get_logger(HW_NAME),
+        "Received linear distance information (L: %f, R: %f)",
+        enc->getTravel(LEFT), enc->getTravel(RIGHT));
+      for (auto i = 0u; i < hw_states_position_.size(); i++)
+      {
+        double delta = linearToAngular(enc->getTravel(i % 2)) - hw_states_position_[i] - hw_states_position_offset_[i];
+
+        // detect suspiciously large readings, possibly from encoder rollover
+        if (std::abs(delta) < 1.0f)
+        {
+          hw_states_position_[i] += delta;
+        }
+        else
+        {
+          // suspicious! drop this measurement and update the offset for subsequent readings
+          hw_states_position_offset_[i] += delta;
+          RCLCPP_WARN(
+            rclcpp::get_logger(HW_NAME),"Dropping overflow measurement from encoder");
+        }
+      }
+    }
+
+    horizon_legacy::Channel<clearpath::DataDifferentialSpeed>::Ptr speed =
+      horizon_legacy::Channel<clearpath::DataDifferentialSpeed>::requestData(polling_timeout_);
+    if (speed)
+    {
+      RCLCPP_DEBUG(
+        rclcpp::get_logger(HW_NAME),
+        "Received linear speed information (L: %f, R: %f)",
+        speed->getLeftSpeed(), speed->getRightSpeed());
+
+      for (auto i = 0u; i < hw_states_velocity_.size(); i++)
+      {
+
+        if (i % 2 == LEFT)
+        {
+          hw_states_velocity_[i] = linearToAngular(speed->getLeftSpeed());
+        }
+        else
+        { // assume RIGHT
+          hw_states_velocity_[i] = linearToAngular(speed->getRightSpeed());
+        }
+      }
+    }
   }
 
+
+hardware_interface::return_type HuskyHardware::configure(
+  const hardware_interface::HardwareInfo & info)
+{
+  if (configure_default(info) != hardware_interface::return_type::OK)
+  {
+    return hardware_interface::return_type::ERROR;
+  }
+
+  RCLCPP_INFO(rclcpp::get_logger(HW_NAME), "Name: %s", info_.name.c_str());
+
+  RCLCPP_INFO(rclcpp::get_logger(HW_NAME), "Number of Joints %u", info_.joints.size());
+
+  hw_start_sec_ = std::stod(info_.hardware_parameters["hw_start_duration_sec"]);
+  hw_stop_sec_ = std::stod(info_.hardware_parameters["hw_stop_duration_sec"]);
+  hw_states_position_.resize(info_.joints.size(), std::numeric_limits<double>::quiet_NaN());
+  hw_states_position_offset_.resize(info_.joints.size(), std::numeric_limits<double>::quiet_NaN());
+  hw_states_velocity_.resize(info_.joints.size(), std::numeric_limits<double>::quiet_NaN());
+  hw_commands_.resize(info_.joints.size(), std::numeric_limits<double>::quiet_NaN());
+
+  wheel_diameter_ = std::stod(info_.hardware_parameters["wheel_diameter"]);
+  max_accel_ = std::stod(info_.hardware_parameters["max_accel"]);
+  max_speed_ = std::stod(info_.hardware_parameters["max_speed"]);
+  polling_timeout_ = std::stod(info_.hardware_parameters["polling_timeout"]);
+
+  serial_port_ = info_.hardware_parameters["serial_port"];
+
+  RCLCPP_INFO(rclcpp::get_logger(HW_NAME), "Port: %s", serial_port_.c_str());
+  horizon_legacy::connect(serial_port_);
+  horizon_legacy::configureLimits(max_speed_, max_accel_);
+  resetTravelOffset();
+
+  for (const hardware_interface::ComponentInfo & joint : info_.joints)
+  {
+    // HuskyHardware has exactly two states and one command interface on each joint
+    if (joint.command_interfaces.size() != 1)
+    {
+      RCLCPP_FATAL(
+        rclcpp::get_logger(HW_NAME),
+        "Joint '%s' has %d command interfaces found. 1 expected.", joint.name.c_str(),
+        joint.command_interfaces.size());
+      return hardware_interface::return_type::ERROR;
+    }
+
+    if (joint.command_interfaces[0].name != hardware_interface::HW_IF_VELOCITY)
+    {
+      RCLCPP_FATAL(
+        rclcpp::get_logger(HW_NAME),
+        "Joint '%s' have %s command interfaces found. '%s' expected.", joint.name.c_str(),
+        joint.command_interfaces[0].name.c_str(), hardware_interface::HW_IF_VELOCITY);
+      return hardware_interface::return_type::ERROR;
+    }
+
+    if (joint.state_interfaces.size() != 2)
+    {
+      RCLCPP_FATAL(
+        rclcpp::get_logger(HW_NAME),
+        "Joint '%s' has %d state interface. 2 expected.", joint.name.c_str(),
+        joint.state_interfaces.size());
+      return hardware_interface::return_type::ERROR;
+    }
+
+    if (joint.state_interfaces[0].name != hardware_interface::HW_IF_POSITION)
+    {
+      RCLCPP_FATAL(
+        rclcpp::get_logger(HW_NAME),
+        "Joint '%s' have '%s' as first state interface. '%s' and '%s' expected.",
+        joint.name.c_str(), joint.state_interfaces[0].name.c_str(),
+        hardware_interface::HW_IF_POSITION);
+      return hardware_interface::return_type::ERROR;
+    }
+
+    if (joint.state_interfaces[1].name != hardware_interface::HW_IF_VELOCITY)
+    {
+      RCLCPP_FATAL(
+        rclcpp::get_logger(HW_NAME),
+        "Joint '%s' have '%s' as second state interface. '%s' expected.", joint.name.c_str(),
+        joint.state_interfaces[1].name.c_str(), hardware_interface::HW_IF_VELOCITY);
+      return hardware_interface::return_type::ERROR;
+    }
+  }
+
+  status_ = hardware_interface::status::CONFIGURED;
+  return hardware_interface::return_type::OK;
+}
+
+std::vector<hardware_interface::StateInterface> HuskyHardware::export_state_interfaces()
+{
+  std::vector<hardware_interface::StateInterface> state_interfaces;
+  for (auto i = 0u; i < info_.joints.size(); i++)
+  {
+    state_interfaces.emplace_back(hardware_interface::StateInterface(
+      info_.joints[i].name, hardware_interface::HW_IF_POSITION, &hw_states_position_[i]));
+    state_interfaces.emplace_back(hardware_interface::StateInterface(
+      info_.joints[i].name, hardware_interface::HW_IF_VELOCITY, &hw_states_velocity_[i]));
+  }
+
+  return state_interfaces;
+}
+
+std::vector<hardware_interface::CommandInterface> HuskyHardware::export_command_interfaces()
+{
+  std::vector<hardware_interface::CommandInterface> command_interfaces;
+
+  for (auto i = 0u; i < info_.joints.size(); i++)
+  {
+    command_interfaces.emplace_back(hardware_interface::CommandInterface(
+      info_.joints[i].name, hardware_interface::HW_IF_VELOCITY, &hw_commands_[i]));
+  }
+
+  return command_interfaces;
+}
+
+hardware_interface::return_type HuskyHardware::start()
+{
+  RCLCPP_INFO(rclcpp::get_logger(HW_NAME), "Starting ...please wait...");
+
+  for (auto i = 0; i <= hw_start_sec_; i++)
+  {
+    rclcpp::sleep_for(std::chrono::seconds(1));
+    RCLCPP_INFO(
+      rclcpp::get_logger(HW_NAME), "%.1f seconds left...", hw_start_sec_ - i);
+  }
+
+  // set some default values
+  for (auto i = 0u; i < hw_states_position_.size(); i++)
+  {
+    if (std::isnan(hw_states_position_[i]))
+    {
+      hw_states_position_[i] = 0;
+      hw_states_position_offset_[i] = 0;
+      hw_states_velocity_[i] = 0;
+      hw_commands_[i] = 0;
+    }
+  }
+
+  status_ = hardware_interface::status::STARTED;
+
+  RCLCPP_INFO(rclcpp::get_logger(HW_NAME), "System Successfully started!");
+
+  return hardware_interface::return_type::OK;
+}
+
+hardware_interface::return_type HuskyHardware::stop()
+{
+  RCLCPP_INFO(rclcpp::get_logger(HW_NAME), "Stopping ...please wait...");
+
+  for (auto i = 0u; i <= hw_stop_sec_; i++)
+  {
+    rclcpp::sleep_for(std::chrono::seconds(1));
+    RCLCPP_INFO(
+      rclcpp::get_logger(HW_NAME), "%.1f seconds left...", hw_stop_sec_ - i);
+  }
+
+  status_ = hardware_interface::status::STOPPED;
+
+  RCLCPP_INFO(rclcpp::get_logger(HW_NAME), "System successfully stopped!");
+
+  return hardware_interface::return_type::OK;
+}
+
+hardware_interface::return_type HuskyHardware::read()
+{
+  RCLCPP_DEBUG(rclcpp::get_logger(HW_NAME), "Reading from hardware");
+
+  updateJointsFromHardware();
+
+  RCLCPP_DEBUG(rclcpp::get_logger(HW_NAME), "Joints successfully read!");
+
+  return hardware_interface::return_type::OK;
+}
+
+hardware_interface::return_type HuskyHardware::write()
+{
+  RCLCPP_DEBUG(rclcpp::get_logger(HW_NAME), "Writing to hardware");
+
+  writeCommandsToHardware();
+
+  RCLCPP_DEBUG(rclcpp::get_logger(HW_NAME), "Joints successfully written!");
+
+  return hardware_interface::return_type::OK;
+}
 
 }  // namespace husky_base
+
+#include "pluginlib/class_list_macros.hpp"
+PLUGINLIB_EXPORT_CLASS(
+  husky_base::HuskyHardware, hardware_interface::SystemInterface)
